+++ src/senaite/core/browser/viewlets/toolbar.py
@@ -78,16 +78,23 @@ class ToolbarViewletManager(OrderedViewletManager):
 
     def get_toolbar_logo(self):
         registry = getUtility(IRegistry)
-        settings = registry.forInterface(
-            ISiteSchema, prefix="senaite", check=False)
         portal_url = self.portal_state.portal_url()
         try:
-            logo = settings.toolbar_logo
+            logo = registry["senaite.toolbar_logo"]
         except AttributeError:
             logo = LOGO
         if not logo:
             logo = LOGO
         return portal_url + logo
+
+    def get_toolbar_styles(self):
+        registry = getUtility(IRegistry)
+        try:
+            styles = registry["senaite.toolbar_logo_styles"]
+        except AttributeError:
+            return "height:15px;"
+        css = map(lambda style: "{}:{};".format(*style), styles.items())
+        return " ".join(css)
 
     def get_lims_setup_url(self):
         portal_url = self.portal_state.portal().absolute_url()